Refactoring folders and adding some layouts for the announcement item.
