+++ app/src/main/java/com/effeta/miparroquiaandroid/common/Status.kt
@@ -1,0 +1,10 @@
+package com.effeta.miparroquiaandroid.common
+
+/**
+ * Created by aulate on 20/2/18.
+ */
+enum class Status {
+    LOADING,
+    ERROR,
+    COMPLETE
+}
