+++ app/src/main/res/layout-v21/item_announcement.xml
@@ -6,6 +6,7 @@
     android:layout_height="wrap_content"
     android:layout_margin="8dp"
     android:background="@drawable/item_announcement"
+    android:elevation="2dp"
     android:orientation="vertical"
     android:paddingBottom="@dimen/activity_horizontal_margin"
     android:paddingEnd="@dimen/activity_vertical_margin"
